+++ hal/ia32/Makefile
@@ -5,7 +5,8 @@
 # Author: Pawel Pisarczyk
 #
 
-OBJS += $(addprefix $(PREFIX_O)hal/ia32/, _init.o _exceptions.o _interrupts.o syspage.o multiboot.o spinlock.o exceptions.o interrupts.o cpu.o pmap.o timer.o hal.o string.o pci.o)
+OBJS += $(addprefix $(PREFIX_O)hal/ia32/, _init.o _exceptions.o _interrupts.o multiboot.o spinlock.o exceptions.o interrupts.o cpu.o pmap.o timer.o hal.o string.o pci.o)
+CFLAGS += -Ihal/ia32
 
 ifeq ($(CONSOLE), vga)
 	OBJS += $(PREFIX_O)hal/ia32/console.o
