+++ hal/ia32/_exceptions.S
@@ -16,7 +16,7 @@
 
 #define __ASSEMBLY__
 
-#include "cpu.h"
+#include <arch/cpu.h>
 
 .text
