+++ hal/ia32/config.h
@@ -1,0 +1,26 @@
+/*
+ * Phoenix-RTOS
+ *
+ * Operating system kernel
+ *
+ * Configuration file for ia2
+ *
+ * Copyright 2021 Phoenix Systems
+ * Author: Hubert Buczynski
+ *
+ * This file is part of Phoenix-RTOS.
+ *
+ * %LICENSE%
+ */
+
+#ifndef _HAL_CONFIG_H_
+#define _HAL_CONFIG_H_
+
+#ifndef __ASSEMBLY__
+
+#include "../../include/arch/syspage-ia32.h"
+
+#endif
+
+
+#endif
