+++ hal/ia32/multiboot.h
@@ -40,6 +40,7 @@
 
 #ifndef __ASSEMBLY__
 
+#include "arch/types.h"
 
 typedef struct {
 	u32 flags;
