+++ hal/ia32/pci.c
@@ -14,9 +14,10 @@
  */
 
 #include "../../include/errno.h"
-#include "cpu.h"
+#include "../cpu.h"
+#include "../spinlock.h"
 #include "pci.h"
-#include "spinlock.h"
+#include "ia32.h"
 
 
 struct {
