+++ hal/ia32/spinlock.c
@@ -14,14 +14,61 @@
  * %LICENSE%
  */
 
-#include "spinlock.h"
-#include "cpu.h"
-
+#include "../spinlock.h"
+#include "../cpu.h"
 
 struct {
 	spinlock_t spinlock;
 	spinlock_t *first;
 } spinlocks;
+
+
+void hal_spinlockSet(spinlock_t *spinlock, spinlock_ctx_t *sc)
+{
+	__asm__ volatile
+	(" \
+		pushf; \
+		popl %%ebx; \
+		cli; \
+	1: \
+		xorl %%eax, %%eax; \
+		xchgl %1, %%eax; \
+		cmp $0, %%eax; \
+		jz 1b; \
+		movl %%ebx, (%0)"
+	:
+	: "r" (sc), "m" (spinlock->lock)
+	: "eax", "ebx", "memory");
+
+	hal_cpuGetCycles((void *)&spinlock->b);
+}
+
+
+void hal_spinlockClear(spinlock_t *spinlock, spinlock_ctx_t *sc)
+{
+	hal_cpuGetCycles((void *)&spinlock->e);
+
+	/* Calculate maximum and minimum lock time */
+	if ((cycles_t)(spinlock->e - spinlock->b) > spinlock->dmax)
+		spinlock->dmax = spinlock->e - spinlock->b;
+
+	if (spinlock->e - spinlock->b < spinlock->dmin)
+		spinlock->dmin = spinlock->e - spinlock->b;
+
+	__asm__ volatile
+	(" \
+		xorl %%eax, %%eax; \
+		incl %%eax; \
+		xchgl %0, %%eax; \
+		movl %1, %%eax; \
+		pushl %%eax; \
+		popf"
+	:
+	: "m" (spinlock->lock), "r" (*sc)
+	: "eax", "memory");
+
+	return;
+}
 
 
 static void _hal_spinlockCreate(spinlock_t *spinlock, const char *name)
