--- hal/ia32/syspage.c
@@ -1,21 +1,0 @@
-/*
- * Phoenix-RTOS
- *
- * Operating system kernel
- *
- * System information page (prepared by kernel loader)
- *
- * Copyright 2012, 2016 Phoenix Systems
- * Copyright 2005 Pawel Pisarczyk
- * Author: Pawel Pisarczyk
- *
- * This file is part of Phoenix-RTOS.
- *
- * %LICENSE%
- */
-
-#include "syspage.h"
-
-
-/* Syspage */
-syspage_t *syspage;
